Moved MultipartHelper.cs to Helper folder.
