+++ WebService/Helper/MultipartHelper.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Linq;
 
-namespace WebService.Services
+namespace WebService.Helper
 {
     public class MultipartHelper
     {
