+++ WebService/Services/MultipartMarshaller.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using WebService.Helper;
 
 namespace WebService.Services
 {
